requirements.txt
python-dotenv
openai
pyupbit
pyyaml
ta
bs4
gitpython
translate
rich